+++ CHANGELOG.md
@@ -381,3 +381,5 @@ Added test case for test action.
 * **2015-10-30** Add three new test cases for unique two step scenarios (add, change, delete).  Fixed data in 
 testgraph.py.  get_step_triples now always returns a Graph. All tests (N=143) pass. Simplify Pump calls in 
 test_vivopump.py
+* **2015-10-31** Three new tests added to `PumpUpdateLiteralsTestCase` for testing unicode add, change and delete. All
+tests pass.